fix verification negative tests about signature
The tests "malformed signature" and "wrong signature" were actually failing because of the expiration time, which was less than `now`, and not because of the signature. This because the expiration time check is performed before the signature check. See: https://github.com/spruceid/siwe/blob/main/packages/siwe/lib/client.ts#L263-L276