Adds test data for RTLO characters
Also adds a convention "STRICT CHECK =<message>" to tell the test runner the ensure the test case throws for the right reason (parser errors are not very descriptive and in some cases we want to check for specific failures).

The STRICT CHECKs for RTLO cases are "commented out" because the APG library chokes on those characters. If this is ever "fixed" upstream, we can uncomment to ensure our RTLO checks will still work.